Working with the concepts of virtual function and how does it affect while working with pointers
Exercise:
Make a pointer out of baseclass,
point it to store the object of derivedclass

Now using the baseclass pointer try to call the
display() function, which is contained by both the
derivedclass and baseclass respectively.

Now, if the baseclass member function is written
without the virtual keyword, the baseclass pointer
will still point to the memberfunction of the
baseclass not to the deriveclass memberfunction

Moment after virtual keyword was added, it was
able to call the display function of the derived
class, still being the pointer of the baseclass.