In this code,
tut45.cpp
We discuss the elements of lists, as to how lists are different from vectors, and how lists are different from arrays.

Addittionally, we discuss that lists advantages and disadvantages while working over other data structures.

In the code, we implemented the basic syntax of the list.
Performed Adding elements into the list, displaying the list using the concept of iterator and then sorting the list.

Sorting capability in contrast to vectors is inbuilt in this data structure, whereas for vectors, we had to
include algorithm.h to perform sort over the vectors.